Added compressed execution

+++ makeBATCH.py
@@ -24,6 +24,9 @@ parser.add_argument("path", help="Path to the input parameters file [input_param
 parser.add_argument("batch_name", nargs="?",
     help="Name for the produced batch file. "\
     "Will use the one defined in the parameter file if ommited.")
+parser.add_argument("-c", "--compact", action="store_true",
+    help="If provided, will reduce the number of subfolder produced " \
+    "by combining executions with similar parameters.")
 
 p = parser.parse_args()
 
@@ -134,65 +137,74 @@ for i,param_vals in enumerate(comb(*param_space),1):
     if 'observer_coordinates' in multival:
         P['observer_coordinates'] = coords
 
-    fold_name = dir_name + \
-        os.sep.join(k+"_%s" % v for k,v in local_params.iteritems()) + os.sep
+    if p.compact:
+        fold_name = dir_name + os.sep.join(
+            "%s_%s" % (k,v) for k,v in local_params.iteritems() \
+            if k in ["observer_coordinates", "wavelength", "layer"]
+        ) + os.sep
+    else:
+        fold_name = dir_name + os.sep.join(
+            "%s_%s" % (k,v) for k,v in local_params.iteritems()
+        ) + os.sep
 
-    os.makedirs(fold_name)
-    #print fold_name
+    unique_ID = '-'.join( "%s_%s" % item for item in local_params.iteritems() )
 
     wavelength = "%03d" % P["wavelength"]
 
-    # Linking files
-    mie_file = "%s_RH%02d_0.%s0um.mie.out" % (
-        params['aerosol_profile'],
-        params['relative_humidity'],
-        wavelength )
-    os.symlink(
-        os.path.abspath(mie_file),
-        fold_name+"aerosol.mie.out" )
+    if not os.path.isdir(fold_name):
+        os.makedirs(fold_name)
 
-    for l,lamp in enumerate(lamps,1):
+        # Linking files
+        mie_file = "%s_RH%02d_0.%s0um.mie.out" % (
+            params['aerosol_profile'],
+            params['relative_humidity'],
+            wavelength )
         os.symlink(
-            os.path.abspath("fctem_wl_%s_lamp_%s.dat" % (wavelength,lamp)),
-            fold_name+exp_name+"_fctem_%03d.dat" % l )
+            os.path.abspath(mie_file),
+            fold_name+"aerosol.mie.out" )
 
-    ppath = os.environ['PATH'].split(os.pathsep)
-    illumpath = filter(lambda s: "illumina" in s and "bin" in s, ppath)[0]
-    os.symlink(
-        os.path.abspath(illumpath+"/illumina"),
-        fold_name+"illumina" )
+        for l,lamp in enumerate(lamps,1):
+            os.symlink(
+                os.path.abspath("fctem_wl_%s_lamp_%s.dat" % (wavelength,lamp)),
+                fold_name+exp_name+"_fctem_%03d.dat" % l )
 
-    # Copying layer data
-    layer = P["layer"]
-
-    obs_fold = os.path.join(
-        "obs_data",
-        coords,
-        str(layer)
-    )
-
-    os.symlink(
-        os.path.abspath(os.path.join(obs_fold,"srtm.bin")),
-        fold_name+exp_name+"_topogra.bin"
-    )
-
-    for name in ["obstd","obsth","obstf","altlp"]:
+        ppath = os.environ['PATH'].split(os.pathsep)
+        illumpath = filter(lambda s: "illumina" in s and "bin" in s, ppath)[0]
         os.symlink(
-            os.path.abspath(os.path.join(obs_fold,"%s_%s.bin" % \
-                ( exp_name, name ) )),
-            fold_name+"%s_%s.bin" % \
-                ( exp_name, name )
+            os.path.abspath(illumpath+"/illumina"),
+            fold_name+"illumina" )
+
+        # Copying layer data
+        layer = P["layer"]
+
+        obs_fold = os.path.join(
+            "obs_data",
+            coords,
+            str(layer)
         )
 
-    for l,lamp in enumerate(lamps,1):
         os.symlink(
-            os.path.abspath(os.path.join(obs_fold,"%s_%s_lumlp_%s.bin" % \
-                ( exp_name, wavelength, lamp ) )),
-            fold_name+"%s_lumlp_%03d.bin" % \
-                ( exp_name, l )
+            os.path.abspath(os.path.join(obs_fold,"srtm.bin")),
+            fold_name+exp_name+"_topogra.bin"
         )
 
-    reflectance = refls[wls.index(P["wavelength"])]
+        for name in ["obstd","obsth","obstf","altlp"]:
+            os.symlink(
+                os.path.abspath(os.path.join(obs_fold,"%s_%s.bin" % \
+                    ( exp_name, name ) )),
+                fold_name+"%s_%s.bin" % \
+                    ( exp_name, name )
+            )
+
+        for l,lamp in enumerate(lamps,1):
+            os.symlink(
+                os.path.abspath(os.path.join(obs_fold,"%s_%s_lumlp_%s.bin" % \
+                    ( exp_name, wavelength, lamp ) )),
+                fold_name+"%s_lumlp_%03d.bin" % \
+                    ( exp_name, l )
+            )
+
+        reflectance = refls[wls.index(P["wavelength"])]
 
     # Create illumina.in
     input_data = (
@@ -237,34 +249,42 @@ for i,param_vals in enumerate(comb(*param_space),1):
         ((ds.pixel_size(layer)/2, "Minimal distance to nearest light source [m]"),)
     )
 
-    with open(fold_name+"illumina.in",'w') as f:
+    with open(fold_name+unique_ID+".in",'w') as f:
         lines = ( input_line(*izip(*line_data)) for line_data in input_data )
         f.write( '\n'.join(lines) )
 
     # Write execute script
-    with open(fold_name+"execute",'w') as f:
-        f.write("#!/bin/sh\n")
-        f.write("#SBATCH --job-name=Illumina\n")
-        f.write("#SBATCH --time=%d:00:00\n" % \
-            params["estimated_computing_time"])
-        f.write("#SBATCH --mem-per-cpu=1920\n")
-        f.write("cd %s\n" % os.path.abspath(fold_name))
-        f.write("umask 0011\n")
+    if not os.path.isfile(fold_name+"execute"):
+        with open(fold_name+"execute",'w') as f:
+            f.write("#!/bin/sh\n")
+            f.write("#SBATCH --job-name=Illumina\n")
+            f.write("#SBATCH --time=%d:00:00\n" % \
+                params["estimated_computing_time"])
+            f.write("#SBATCH --mem-per-cpu=1920\n")
+            f.write("cd %s\n" % os.path.abspath(fold_name))
+            f.write("umask 0011\n")
+        os.chmod(fold_name+"execute",0o777)
+
+        # Append execution to batch list
+        with open(
+            p.path + \
+                '/' + \
+                params['batch_file_name'] + \
+                "_%d" % ((count/300)+1) ,
+            'a' ) as f:
+            f.write("cd %s\n" % os.path.abspath(fold_name))
+            f.write("sbatch ./execute\n")
+            f.write("sleep 0.05\n")
+
+        count += 1
+
+    # Add current parameters execution to execution script
+    with open(fold_name+"execute",'a') as f:
+        f.write("cp %s.in illumina.in\n" % unique_ID)
         f.write("./illumina\n")
-    os.chmod(fold_name+"execute",0o777)
-
-    # Append execution to batch list
-    with open(
-        sys.argv[1] + \
-            '/' + \
-            params['batch_file_name'] + \
-            "_%d" % ((count/300)+1) ,
-        'a' ) as f:
-        f.write("cd %s\n" % os.path.abspath(fold_name))
-        f.write("sbatch ./execute\n")
-        f.write("sleep 0.05\n")
-
-    count += 1
+        f.write("mv %s.out %s_%s.out\n" % (exp_name,exp_name,unique_ID))
+        f.write("mv %s_pcl.bin %s_pcl_%s.bin\n" % (exp_name,exp_name,unique_ID))
+        f.write("mv %s_pcw.bin %s_pcw_%s.bin\n" % (exp_name,exp_name,unique_ID))
 
 print "Final count:", count
